flask/post.py/requirements.txt
fastapi==0.95.0
tortoise-orm==0.21.0
uvicorn==0.22.0
pydantic==1.10.0

